requirements.txt
langchain==0.3.0
langchain-community
langchain-chroma
langchain-openai
chromadb
gitpython
python-dotenv
tiktoken